pyproject.toml
[tool.poetry]
name = "kittens-exhibition"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
